# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "简易个人订阅制软件管理平台 API",
        "title": "Nexus Core API",
        "contact": {},
        "version": "0.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/products": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Create a product",
                "parameters": [
                    {
                        "description": "Create Product",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateProductCommand"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.CommonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.CommonResponse"
                        }
                    }
                }
            }
        },
...
        "/products/versions": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Create a product version",
                "parameters": [
                    {
                        "description": "Create Product Version",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateProductVersionCommand"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Version"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.CommonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.CommonResponse"
                        }
                    }
                }
            }
        },
        "/products/versions/deprecate": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Deprecate a version",
                "parameters": [
                    {
                        "description": "Deprecate Version",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DeprecateVersionCommand"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.CommonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.CommonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.CommonResponse"
                        }
                    }
                }
            }
        },
        "/products/versions/release": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Release a new version",
                "parameters": [
                    {
                        "description": "Release New Version",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ReleaseNewVersionCommand"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Version"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get product by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.CommonResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.CommonResponse"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Delete product",
                "parameters": [
                    {
                        "description": "{\\",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.CommonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.CommonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.CommonResponse"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Update product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Product",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateProductCommand"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.CommonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.CommonResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.CommonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.CommonResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.CommonResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据，可选"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
...
        "dto.CreateProductCommand": {
            "description": "Command to create a product",
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "description": "产品描述",
                    "type": "string"
                },
                "name": {
                    "description": "产品名称",
                    "type": "string"
                }
            }
        },
        "dto.CreateProductVersionCommand": {
            "description": "Product version DTO",
            "type": "object",
            "required": [
                "product_id",
                "version_code"
            ],
            "properties": {
                "description": {
                    "description": "版本描述",
                    "type": "string"
                },
                "product_id": {
                    "description": "所属产品ID",
                    "type": "integer"
                },
                "release_date": {
                    "description": "发布时间",
                    "type": "string"
                },
                "release_method": {
                    "description": "0立即发布，1定时发布，2暂不发布",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.ReleaseMethod"
                        }
                    ]
                },
                "version_code": {
                    "description": "版本号",
                    "type": "string"
                }
            }
        },
        "dto.DeprecateVersionCommand": {
            "type": "object",
            "required": [
                "product_id",
                "version_id"
            ],
            "properties": {
                "product_id": {
                    "type": "integer"
                },
                "version_id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.ReleaseMethod": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-comments": {
                "ReleaseHold": "2 暂不发布",
                "ReleaseImmediate": "0 立即发布",
                "ReleaseScheduled": "1 定时发布"
            },
            "x-enum-descriptions": [
                "0 立即发布",
                "1 定时发布",
                "2 暂不发布"
            ],
            "x-enum-varnames": [
                "ReleaseImmediate",
                "ReleaseScheduled",
                "ReleaseHold"
            ]
        },
        "dto.ReleaseNewVersionCommand": {
            "type": "object",
            "required": [
                "product_id",
                "version_id"
            ],
            "properties": {
                "product_id": {
                    "description": "所属产品ID",
                    "type": "integer"
                },
                "release_date": {
                    "description": "发布时间",
                    "type": "string"
                },
                "version_id": {
                    "description": "版本ID",
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdateProductCommand": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "entity.Product": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "产品详细描述",
                    "type": "string"
                },
                "id": {
                    "description": "产品唯一标识符",
                    "type": "integer"
                },
                "minSupportedVersionID": {
                    "description": "最低支持的版本ID，用于版本兼容性检查，按时间，在此之前发布的版本将无法使用",
                    "type": "integer"
                },
                "name": {
                    "description": "产品名称，具有唯一性",
                    "type": "string"
                },
                "versionList": {
                    "description": "产品版本列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Version"
                    }
                }
            }
        },
        "entity.Version": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "版本详细说明",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "releaseDate": {
                    "description": "版本发布时间",
                    "type": "string"
                },
                "status": {
                    "description": "版本状态，用于标识版本是否可用",
                    "type": "integer"
                },
                "versionCode": {
                    "description": "版本号，遵循语义化版本规范",
                    "type": "string"
                }
            }
        }
    }
}